I removed the two split out test scripts and converted them back to the one test script: execution.events.js being sure to convert the test code to RTL from Enzyme.
