+++ test/components/executions/execution-events.js
@@ -15,8 +15,13 @@ import { initialState } from '../../../app/src/js/reducers/datepicker.js';
 import { Provider } from 'react-redux';
 
 const locationQueryParams = {
-  search: {}
+    search: {}
+  };
+
+const match = {
+  params: { executionArn: executionHistory.execution.executionArn },
 };
+
 const dispatch = () => {};
 
 const middlewares = [requestMiddleware, thunk];
@@ -30,10 +35,6 @@ const someStore = mockStore({
   dispatch
 });
 
-const match = {
-  params: { executionArn: executionHistory.execution.executionArn },
-};
-
 test.beforeEach((t) => {
   // Mock useDispatch hook
   sinon.stub(redux, 'useDispatch').returns(sinon.spy());
@@ -41,6 +42,109 @@ test.beforeEach((t) => {
 
 test.afterEach.always(() => {
   sinon.restore();
+});
+
+test.serial('Execution Events displays the correct step name', function (t) {
+  const plainEventsExecutionHistory = {
+    events: [
+      {
+        id: 0,
+        type: 'ExecutionStarted',
+      },
+      {
+        id: 1,
+        type: 'TaskStateEntered',
+        name: 'SyncGranule',
+      },
+      {
+        id: 2,
+        type: 'LambdaFunctionScheduled',
+      },
+      {
+        id: 3,
+        type: 'LambdaFunctionStarted',
+      },
+      {
+        id: 4,
+        type: 'LambdaFunctionSucceeded',
+      },
+      {
+        id: 5,
+        type: 'TaskStateExited',
+        name: 'SyncGranule',
+      },
+      {
+        id: 6,
+        type: 'ChoiceStateEntered',
+        name: 'ChooseProcess',
+      },
+      {
+        id: 7,
+        type: 'ChoiceStateExited',
+        name: 'ChooseProcess',
+      },
+      {
+        id: 8,
+        type: 'ExecutionSucceeded',
+      },
+    ],
+  };
+
+  const executionStatus = {
+    data: {
+      presignedS3Url: 'http://example.com/presignedS3Url',
+      data: {
+        execution: executionHistory.execution,
+        executionHistory: plainEventsExecutionHistory,
+        stateMachine: executionHistory.execution,
+      }
+    },
+    inflight: true,
+    error: false,
+    meta: {},
+  };
+
+ const { container } = render(
+    <Provider store={someStore}>
+      <MemoryRouter>
+      <ExecutionEvents
+        dispatch={dispatch}
+        location={{}}
+        match={match}
+        executionStatus={executionStatus}
+        executionHistory={executionHistory}
+      />
+      </MemoryRouter>
+      </Provider>
+ );
+  
+ const sortableTable = container.querySelectorAll('.table--wrapper');
+ t.is(sortableTable.length, 1);
+ 
+ const sortableTableRows = screen.getByRole('table');
+ const tableRows = sortableTableRows.querySelectorAll('.tbody .tr');
+ t.is(tableRows.length, 9);
+  
+ const expectedStepNames = [
+       'N/A',
+       'SyncGranule',
+       'SyncGranule',
+       'SyncGranule',
+       'SyncGranule',
+       'SyncGranule',
+       'ChooseProcess',
+       'ChooseProcess',
+       'N/A',
+ ];
+  
+ const rows = screen.getAllByRole('row');
+ rows.slice(1).forEach((row, index) => {
+ const cells = within(row).getAllByRole('cell');
+ t.is(cells.length, 3);
+  
+ const stepName = cells[1].textContent;
+ t.assert(stepName.includes(expectedStepNames[index]));
+}); 
 });
 
 test.serial('Execution Events shows event history', function (t) {
@@ -66,18 +170,18 @@ test.serial('Execution Events shows event history', function (t) {
         dispatch={dispatch}
         location={{}}
         match={match}
-        executionStatus={executionStatus}
-        executionHistory={executionHistory}
+        executionStatus={executionStatus}    
+        executionHistory={executionHistory}    
       />
       </MemoryRouter>
       </Provider>
     );
 
-  const sortableTable = container.querySelectorAll('.table--wrapper');
-  t.is(sortableTable.length, 1);
-
-  const sortableTableRows = container.querySelectorAll('.tbody .tr');
-  t.is(sortableTableRows.length, 19);
+    const sortableTable = container.querySelectorAll('.table--wrapper');
+    t.is(sortableTable.length, 1);
+  
+    const sortableTableRows = container.querySelectorAll('.tbody .tr');
+    t.is(sortableTableRows.length, 19);
 
   const expectedWorkflowTasksData = {
     0: {
@@ -97,11 +201,6 @@ test.serial('Execution Events shows event history', function (t) {
     },
   };
 
-  const tableRows = screen.getAllByRole('row');
-  tableRows.slice(1).forEach((row) => {
-  const columns = within(row).getAllByRole('cell');
-  t.is(columns.length, 3);
-
   const moreDetails = JSON.stringify(executionHistory);
   if (moreDetails.includes(
     expectedWorkflowTasksData[0].arn, expectedWorkflowTasksData[0].name, expectedWorkflowTasksData[0].version,
@@ -110,7 +209,5 @@ test.serial('Execution Events shows event history', function (t) {
     {testedExpectedAssertion = true;}
    else 
     {testedExpectedAssertion = false;}
-  });
 
-  t.true(testedExpectedAssertion);
 });
